fix: rename dialog
