+++ applications/salk-portal/src/components/AboutDialog.tsx
