+++ applications/salk-portal/src/components/ExplorerSidebar.jsx
@@ -23,7 +23,7 @@ import COMMUNITY from "../assets/images/icons/community.svg";
 import HELP from "../assets/images/icons/help.svg";
 import UP_ICON from "../assets/images/icons/up.svg";
 import { EXPERIMENTS_HASH, SALK_TEAM, ACME_TEAM, COMMUNITY_HASH, SHARED_HASH } from "../constants";
-import AboutCoordAtlasDialog from './AboutCoordAtlasDialog';
+import AboutDialog from './AboutDialog';
 
 const useStyles = makeStyles({
   sidebar: {
@@ -255,7 +255,7 @@ const Sidebar = (props) => {
           <ListItemText primary="Help Center" />
         </ListItemLink>
       </List>
-      <AboutCoordAtlasDialog open={openDialog} handleClose={handleDialogToggle}></AboutCoordAtlasDialog>
+      <AboutDialog open={openDialog} handleClose={handleDialogToggle}></AboutDialog>
       <Box className={classes.footer}>
         <Typography>Funded by NIH</Typography>
         <Button disableRipple variant="text" onClick={handleDialogToggle}>Learn More</Button>
